big refactoring; fixes and new features
- change HighsDomain so that multiple HighsCutPool instances can be
registeres for propagation. HighsCutPool automatically notifies all
registered domains when new cuts are added. This allows for easy
creation of child instances of HighsDomain and is also helpful in
preparation for a parallel tree search so that subproblem dives can
store conflicts in a local cutpool independently of other parallel tree
search tasks.

- implement extraction of cliques from the objective function whenever a
new improving solution is found. This allows to the clique table to
derive fixings and substitutions from new conflict graph edges coming
from the new cutoff bound.

- big refactoring of primal heuristics. Added a class HighsPrimalHeuristics which
provides member functions to run different heuristics and collects
statistics. Added several heuristics: A randomized round-and-propagate heuristic,
feasibility pump, and central rounding. When randomized rounding did not find a
solution and RENS also failed, then furst try the central rounding heuristics and
if still no solution is found try feasibility pump. Central rounding
computes the analytic center with IPX and then searches on the line
segment connecting the first lp solution to the analytic center for a
feasible integer point in proximity with a round-and-propagate scheme.

- implemented reduced cost fixing.