add robin hood hashtable; major performance improvement for clique table
Add a hashtable implementation that uses robin hood hashing. Use new hash table for cliques of size two
to speed up queries of edges for the case of many size two cliques from probing. The cliques of size two
are stored in a seprate set of cliques for the inverted index.