Fix EOF detection in LP file reader
The condition `this->linebufferpos == this->linebuffer.size()` would never be `true` except potentially (but not necessarily) at the end of the file, since whenever an `'\n'` is encountered, `this->linebufferpos` is manually set to correspond to the end of the line (on line 998).

Moreover, the condition would mean that an actual EOF would not be properly handled when the final line of the file is not empty, which would instead cause an infinite loop. In particular, this infinite loop is encountered for the example given in gh-840:

```
4d 69 6e 69 6d 69 7a 65 0a 20 78 30 20 2b 20 78 31 0a 73 75 62 6a 65 63 74 20 74 6f 0a 20 63 31 3a 20 78 30 20 2b 20 78 31 20 3c 3d 20 31 0a 45 6e 64
```

This closes gh-840.