ENH: Add meson build system
This is almost at parity with the `cmake` build system, and should be
simpler to maintain. It is also easier to make `shared` and `static`
libraries from the commandline / at the subproject level. The only
missing feature is that `csharp` is not a supported language so the
interface cannot be built with `meson`.