Minor correction to nqueens.py
Preparing to use this for a demo, I noticed a minor inconsistency.

The `axis` kwarg specifies the axis over which to sum, with 0 being rows and 1 being columns. But to get row totals you have to sum over columns, and vice versa.

Although the symmetry in this problem means it makes no difference, I thought it'd be better to avoid confusion.